Splits large TextIOTest into TextIOReadTest and TextIOWriteTest
